Add transaction-details component
