+++ ui/app/components/ui/popover/popover.stories.js
@@ -2,6 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { storiesOf } from '@storybook/react'
 import PopOver from './popover.component'
+import TransactionDetails from '../transaction-details'
 import { text } from '@storybook/addon-knobs/react'
 
 import { Tab, Tabs, TabList, TabPanel } from 'react-tabs'
@@ -82,41 +83,6 @@ const Activity = () => (
     </ul>
   </div>
 )
-
-const TransactionDetails = ({
-  amount,
-  fee,
-  ethCost,
-}) => (
-  <div className="details">
-    <div className="details-amounts">
-      <div className="col">
-        Amount<br />
-        Transaction Fee
-      </div>
-      <div className="col">
-        ${amount}<br />
-        ${fee}
-      </div>
-    </div>
-    <hr />
-    <div className="details-totals">
-      <div className="col">
-        <h3>Total amount</h3>
-      </div>
-      <div className="col">
-        <h3>${amount + fee}</h3>
-        <p>{ethCost} ETH</p>
-      </div>
-    </div>
-  </div>
-)
-
-TransactionDetails.propTypes = {
-  amount: PropTypes.number.isRequired,
-  fee: PropTypes.number.isRequired,
-  ethCost: PropTypes.number.isRequired,
-}
 
 const Chart = () => (
   <div className="chart"></div>
