+++ ui/app/components/ui/transaction-details/index.js
@@ -1,0 +1,2 @@
+import TransactionDetails from './transaction-details.component'
+export default TransactionDetails
